Refactor: Moved images and footer files out of their folders and updated paths
Moved all image and footer files out of the 'assets' and 'footer' folders to the root directory and deleted the 'assets' and 'footer' folders. Updated all instances of '/assets/...' and '/footer/...' to the respective root paths in the codebase. This refactor was done to test proper publishing on GitHub Pages.